Created Dockerfile and restructured project; Improved README.md
